+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.0.1
+- Teleport to Owner during Journey if they get too far away
 
 # 1.0.0
 - First non-beta release